income-platform/src/broker-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.5
pydantic-settings==2.6.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
httpx==0.27.0
python-dotenv==1.0.1
PyJWT==2.8.0
